aggiunta sicurezza

+++ admin/cancella.php
@@ -2,6 +2,13 @@
 
 include_once '../inc/db.config.php';
 
+session_start();
+if($_SESSION['login'] == false)
+{
+    header('location: index.php');
+    exit;
+}
+
 $id = $_GET['idpratiche'];
 
 $sql = "DELETE FROM pratiche WHERE id_pratica=?";
